pyproject.toml
[project]
name = "expense-tracker-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.0",
    "psycopg2-binary>=2.9.0",
]

[tool.setuptools]
py-modules = ["dbConnection", "main"]
